requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.3
email-validator==2.0.0
WTForms==3.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==20.1.0    
requests==2.31.0
